refactor(schools): absorb the classroom modals from the classrooms pseudo-module
