+++ resources/js/pages/schools/components/create-classroom.tsx
