+++ resources/js/pages/schools/components/delete-classroom-confirmation.tsx
